Buffer period

+++ classes/tank.js
@@ -3,13 +3,12 @@ class Tank extends ScreenObjects{
         super(tankOptions);
         this.amITurningRight = false;
         this.amITurningLeft = false;
-        this.randomID = 'tank' + Math.floor( Math.random() * 1000 );
+        this.randomID = 'tank' + randomNumberGenerator(1000);
         this.rateOfFireBoolean = false;
         this.turretAngle = 0;
         this.configObj = {
             class: tankOptions.class || 'tankSquare', 
             id: this.randomID,
-            src: this.img,
             css: { 
                 'top': this.yPosition+'px', 
                 left: this.xPosition+'px', 
@@ -47,7 +46,8 @@ class Tank extends ScreenObjects{
             xPosition: this.xPosition + parseFloat( this.selector.css('transform-origin').split(' ')[ 0 ] ) + ( Math.sin( (this.turretAngle + this.angleOfDirection) * radiansConversionFactor ) * ( this.hitBox.width / 1.1 ) ), 
             yPosition: this.yPosition + parseFloat( this.selector.css('transform-origin').split(' ')[ 1 ] ) - ( Math.cos( (this.turretAngle + this.angleOfDirection) * radiansConversionFactor ) * ( this.hitBox.height / 1.1 ) ), 
             angleOfDirection: this.angleOfDirection + this.turretAngle,
-            currentGameBoard: this.currentGameBoard
+            currentGameBoard: this.currentGameBoard,
+            tank: this
         });
         setTimeout(() => {
             this.rateOfFireBoolean = false;
@@ -61,6 +61,12 @@ class Tank extends ScreenObjects{
         this.amITurningLeft = false;
     }
     turnLeft(){
+        if( this.angleOfDirection > 360 ){
+            this.angleOfDirection = 0;
+        }
+        if(  0 > this.angleOfDirection ){
+            this.angleOfDirection = 360;
+        }
         this.angleOfDirection -= turnRadius;
         this.turretAngle += turnRadius;
         this.tankTurret.css( 'transform', `rotate(${this.turretAngle}deg)` );
@@ -74,6 +80,12 @@ class Tank extends ScreenObjects{
         this.amITurningRight = false;
     }
     turnRight(){
+        if( this.angleOfDirection > 360 ){
+            this.angleOfDirection = 0;
+        }
+        if(  0 > this.angleOfDirection ){
+            this.angleOfDirection = 360;
+        }
         this.angleOfDirection += turnRadius;
         this.turretAngle -= turnRadius;
         this.tankTurret.css( 'transform', `rotate(${this.turretAngle}deg)` );
@@ -100,6 +112,9 @@ class Tank extends ScreenObjects{
                 this.hitBox.right < cannonBallArray[ collisionIndex ].hitBox.left  ){
             }
             else{
+                if( cannonBallArray[ collisionIndex ].ownerTank === this ){
+                    return;
+                }
                 cannonBallArray[ collisionIndex ].destroyCannonBall();
                 this.destroyTank();
                 //soundsObj.tankDeath.play();
